+++ internal/api/v1/authenticate/response.go
@@ -1,0 +1,7 @@
+package authenticate
+
+import "github.com/Alexander1000/service-auth/internal/model"
+
+type response struct {
+	Result model.Token `json:"result"`
+}