+++ internal/api/v1/authenticate/storage.go
@@ -6,6 +6,6 @@ import (
 )
 
 type storageRepository interface {
-	Authenticate(ctx context.Context, cred model.Credential, pass string) error
+	Authenticate(ctx context.Context, cred model.Credential, pass string) (*model.Token, error)
 }